+++ packages/frontend/src/components/common/Authorization/context.ts
@@ -1,5 +1,0 @@
-import { Ability } from '@casl/ability';
-import { type PossibleAbilities } from '@lightdash/common';
-import { createContext } from 'react';
-
-export const AbilityContext = createContext(new Ability<PossibleAbilities>());