+++ packages/frontend/src/components/common/Table/Table.styles.ts
@@ -135,18 +135,20 @@ export const Table = styled.table<{ $showFooter?: boolean }>`
     }
 
     .sticky-column {
-        position: sticky !important;
         left: 1px;
         z-index: 1;
-        background-color: white !important;
         word-break: break-word;
         :hover {
             white-space: normal;
             background-color: white;
         }
     }
+    th.sticky-column {
+        background: white !important;
+    }
+
     .last-sticky-column {
-        border-right: 2px solid darkgray;
+        border-right: 1.4px solid rgb(189, 189, 189);
     }
 `;