+++ packages/frontend/src/ee/features/embed/SettingsEmbed/EmbedCodeSnippet.tsx
@@ -21,6 +21,7 @@ const projectUuid = '{{projectUuid}}';
 const data = {
     content: {
         type: 'dashboard',
+        projectUuid: projectUuid,
         dashboardUuid: '{{dashboardUuid}}',
         dashboardFiltersInteractivity: {
             enabled: "{{dashboardFiltersInteractivityEnabled}}",
@@ -51,6 +52,7 @@ data = {
     "iat": datetime.datetime.now(tz=datetime.timezone.utc),
     "content": {
         "type": "dashboard",
+        "projectUuid": projectUuid,
         "dashboardUuid": "{{dashboardUuid}}",
         "dashboardFiltersInteractivity": {
             "enabled": "{{dashboardFiltersInteractivityEnabledPython}}",
@@ -68,7 +70,7 @@ data = {
     "userAttributes": {{userAttributes}},
 };
 token = jwt.encode(data, key, algorithm="HS256")
-url = f"{{siteUrl}}/embed/{projectUuid}#{token}"  
+url = f"{{siteUrl}}/embed/{projectUuid}#{token}"
 `,
     [SnippetLanguage.GO]: `
 package main
@@ -91,6 +93,7 @@ func main() {
     type CustomClaims struct {
         Content struct {
             Type                       string \`json:"type"\`
+            ProjectUuid                string \`json:"projectUuid"\`
             DashboardUuid              string \`json:"dashboardUuid"\`
             DashboardFiltersInteractivity struct {
                 Enabled string \`json:"enabled"\`
@@ -113,6 +116,7 @@ func main() {
     claims := CustomClaims{
         Content: struct {
             Type                       string \`json:"type"\`
+            ProjectUuid                string \`json:"projectUuid"\`
             DashboardUuid              string \`json:"dashboardUuid"\`
             DashboardFiltersInteractivity struct {
                 Enabled string \`json:"enabled"\`
@@ -124,6 +128,7 @@ func main() {
             CanDateZoom bool \`json:"canDateZoom"\`
         }{
             Type:          "dashboard",
+            ProjectUuid:   projectUuid,
             DashboardUuid: "{{dashboardUuid}}",
             DashboardFiltersInteractivity: struct {
                 Enabled string \`json:"enabled"\`