+++ packages/frontend/src/providers/Fullscreen/types.ts
@@ -1,0 +1,9 @@
+type FullscreenContextType =
+    | undefined
+    | {
+          enabled: boolean;
+          isFullscreen: boolean;
+          toggleFullscreen: (nextValue?: boolean) => void;
+      };
+
+export default FullscreenContextType;