+++ packages/frontend/src/testing/testUtils.tsx
@@ -12,7 +12,7 @@ import { type FC, type PropsWithChildren, type ReactElement } from 'react';
 import { vi } from 'vitest';
 import MantineProvider from '../providers/MantineProvider';
 import ReactQueryProvider from '../providers/ReactQuery/ReactQueryProvider';
-import { TrackingProvider } from '../providers/Tracking/TrackingProvider';
+import TrackingProvider from '../providers/Tracking/TrackingProvider';
 import AppProviderMock, {
     type AppProviderMockProps,
 } from './__mocks__/providers/AppProvider.mock';