+++ packages/sdk/src/index.ts
@@ -1,0 +1,2 @@
+import Lightdash from './Lightdash';
+export default Lightdash;